feat: documents create & upsert with options
