+++ Sources/Typesense/Models/DocumentUpsertParameters.swift
@@ -1,0 +1,17 @@
+import Foundation
+
+
+
+public struct DocumentUpsertParameters: Codable {
+
+    public var dirtyValues: DirtyValues?
+
+    public init(dirtyValues: DirtyValues? = nil) {
+        self.dirtyValues = dirtyValues
+    }
+
+    public enum CodingKeys: String, CodingKey {
+        case dirtyValues = "dirty_values"
+    }
+
+}
